reorganise screens and their navigators
