+++ components/profileScreen/EditCompanyProfile.tsx
@@ -6,7 +6,7 @@ import { View, Text } from "../Themed";
 import { StyleSheet } from "react-native";
 import Colors from "../../constants/Colors";
 import { TextInput } from "../TextInput";
-import { EditStatus } from "../../screens/EditProfileScreen";
+import { EditStatus } from "../../screens/profile/EditProfileScreen";
 import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
 
 type EditCompanyProfileProps = {
