+++ screens/auth/ResetPasswordScreen.tsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useState } from 'react';
 import { Image, ActivityIndicator, StyleSheet, Pressable } from 'react-native';
 
-import { Text, View } from '../components/Themed';
-import { TextInput } from '../components/TextInput';
+import { Text, View } from '../../components/Themed';
+import { TextInput } from '../../components/TextInput';
 
-import { ArkadButton } from '../components/Buttons';
-import { ArkadText } from '../components/StyledText';
+import { ArkadButton } from '../../components/Buttons';
+import { ArkadText } from '../../components/StyledText';
 
-import { API } from '../api'
+import { API } from '../../api'
 import { StackNavigationProp } from '@react-navigation/stack';
-import { AuthStackParamList } from '../navigation';
+import { AuthStackParamList } from "./AuthNavigator";
 
 type ResetPasswordScreenParams = {
   navigation: StackNavigationProp<
