+++ screens/companies/CompaniesScreen.tsx
@@ -2,14 +2,14 @@ import React, { useEffect, useState } from 'react';
 import { FlatList, StyleSheet, TextInput } from 'react-native';
 import { StackNavigationProp } from '@react-navigation/stack';
 
-import { Text, View } from '../components/Themed';
+import { Text, View } from '../../components/Themed';
 
-import { API } from '../api';
-import { PublicCompanyDto } from '../api/companies';
-import { CompanyListItem } from '../components/companies/CompanyListItem';
-import { CompanyStackParamList } from "../navigation/CompaniesNavigator";
-import ScreenActivityIndicator from '../components/ScreenActivityIndicator';
-import Colors from '../constants/Colors';
+import { API } from '../../api';
+import { PublicCompanyDto } from '../../api/companies';
+import { CompanyListItem } from '../../components/companies/CompanyListItem';
+import { CompanyStackParamList } from "./CompaniesNavigator";
+import ScreenActivityIndicator from '../../components/ScreenActivityIndicator';
+import Colors from '../../constants/Colors';
 
 type companiesNavigation = {
   navigation: StackNavigationProp<
