+++ screens/event/EventDetailsScreen.tsx
@@ -6,21 +6,21 @@ import {
   MaterialCommunityIcons,
 } from "@expo/vector-icons";
 
-import Colors from "../constants/Colors";
+import Colors from "../../constants/Colors";
 
-import { API } from "../api";
-import { bookedEvent, Event } from "../api/events";
+import { API } from "../../api";
+import { bookedEvent, Event } from "../../api/events";
 import {
   CreateTicketDto,
   getTicketForEvent,
   removeTicket,
   Ticket,
-} from "../api/tickets";
+} from "../../api/tickets";
 
-import { View } from "../components/Themed";
-import ScreenActivityIndicator from "../components/ScreenActivityIndicator";
-import { ArkadButton } from "../components/Buttons";
-import { ArkadText } from "../components/StyledText";
+import { View } from "../../components/Themed";
+import ScreenActivityIndicator from "../../components/ScreenActivityIndicator";
+import { ArkadButton } from "../../components/Buttons";
+import { ArkadText } from "../../components/StyledText";
 import QRCode from "react-native-qrcode-svg";
 
 type EventDetailsScreenParams = {
