+++ screens/maps/MapScreen.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { StackNavigationProp } from '@react-navigation/stack';
 import { StyleSheet, ScrollView } from 'react-native';
-import { MapPreview } from '../components/maps/MapPreview';
-import { Map, EMap, FairMap, KarhusetMap, SC1Map, SC2Map } from '../components/maps/MapProps';
-import { View } from '../components/Themed';
-import { MapStackParamList } from "../navigation/MapNavigator";
+import { MapPreview } from '../../components/maps/MapPreview';
+import { Map, EMap, FairMap, KarhusetMap, SC1Map, SC2Map } from '../../components/maps/MapProps';
+import { View } from '../../components/Themed';
+import { MapStackParamList } from "./MapNavigator";
 
 export type mapNavigation = {
   navigation: StackNavigationProp<MapStackParamList, 'MapScreen'>;
