+++ screens/maps/ZoomMapScreen.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { StyleSheet } from 'react-native';
 import { ImageViewer } from 'react-native-image-zoom-viewer';
-import { Map } from '../components/maps/MapProps';
-import ScreenActivityIndicator from '../components/ScreenActivityIndicator';
-import { ArkadText } from '../components/StyledText';
-import { View } from '../components/Themed';
-import Colors from '../constants/Colors';
+import { Map } from '../../components/maps/MapProps';
+import ScreenActivityIndicator from '../../components/ScreenActivityIndicator';
+import { ArkadText } from '../../components/StyledText';
+import { View } from '../../components/Themed';
+import Colors from '../../constants/Colors';
 
 type MapScreenParams = {
   route: {
