+++ screens/studentSessions/templates/SSsApplicationSreen.tsx
@@ -1,16 +1,16 @@
 import React, { useState } from "react";
 import { ScrollView, StyleSheet, TextInput } from "react-native";
 
-import Colors from "../constants/Colors";
+import Colors from "../../../constants/Colors";
 
-import { API } from "../api";
-import { } from "../api/studentsessions";
+import { API } from "../../../api";
+import { } from "../../../api/studentsessions";
 
-import { View } from "../components/Themed";
-import ScreenActivityIndicator from "../components/ScreenActivityIndicator";
-import { ArkadButton } from "../components/Buttons";
-import { ArkadText } from "../components/StyledText";
-import { Role } from "../api/users";
+import { View } from "../../../components/Themed";
+import ScreenActivityIndicator from "../../../components/ScreenActivityIndicator";
+import { ArkadButton } from "../../../components/Buttons";
+import { ArkadText } from "../../../components/StyledText";
+import { Role } from "../../../api/users";
 
 export default function SSsApplicationScreen(companyId: number) {
