+++ screens/studentSessions/templates/SSsApplicationsListScreen.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { StyleSheet } from 'react-native';
 
-import { View } from '../components/Themed';
+import { View } from '../../../components/Themed';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { API } from '../api';
-import { ApplicationsList } from '../components/sSApplication/SSApplicationList';
-import { SSsStackParamlist } from "../navigation/SSsCRepNavigator";
+import { API } from '../../../api';
+import { ApplicationsList } from '../../../components/sSApplication/SSApplicationList';
+import { SSsStackParamlist } from "../SSsCRepNavigator";
 import { ScrollView } from 'react-native-gesture-handler';
-import ScreenActivityIndicator from '../components/ScreenActivityIndicator';
-import { SSApplicationDto } from '../api/sSApplications';
+import ScreenActivityIndicator from '../../../components/ScreenActivityIndicator';
+import { SSApplicationDto } from '../../../api/sSApplications';
 import { useFocusEffect } from '@react-navigation/native';
 import { useCallback } from 'react';
